Sends Map of transcriptMessages as prop to Component
