+++ react/features/transcription/middleware.js
@@ -69,18 +69,18 @@ function _endpointMessageReceived({ dispatch, getState }, next, action) {
             // we add it to the state along with the name of the participant
             // who said given text
             if (!getState()['features/transcription']
-                .transcriptMessages.get(transcriptMessageID)) {
+                .transcriptMessages.has(transcriptMessageID)) {
                 dispatch(addTranscriptMessage(transcriptMessageID,
                     participantName));
             }
+            const { transcriptMessages }
+                = getState()['features/transcription'];
+            const newTranscriptMessage
+                = transcriptMessages.get(transcriptMessageID);
 
             // If this is final result, update the state as a final result
             // and start a count down to remove the subtitle from the state
             if (!isInterim) {
-                const { transcriptMessages }
-                    = getState()['features/transcription'];
-                const newTranscriptMessage
-                    = transcriptMessages.get(transcriptMessageID);
 
                 newTranscriptMessage.final = text;
                 dispatch(updateTranscriptMessage(transcriptMessageID,
@@ -94,10 +94,6 @@ function _endpointMessageReceived({ dispatch, getState }, next, action) {
                 // If the message has a high stability, we can update the
                 // stable field of the state and remove the previously
                 // unstable results
-                const { transcriptMessages }
-                    = getState()['features/transcription'];
-                const newTranscriptMessage
-                    = transcriptMessages.get(transcriptMessageID);
 
                 newTranscriptMessage.stable = text;
                 newTranscriptMessage.unstable = undefined;
@@ -107,11 +103,6 @@ function _endpointMessageReceived({ dispatch, getState }, next, action) {
                 // Otherwise, this result has an unstable result, which we
                 // add to the state. The unstable result will be localed at
                 // the end of the String, after the stable part.
-
-                const { transcriptMessages }
-                    = getState()['features/transcription'];
-                const newTranscriptMessage
-                    = transcriptMessages.get(transcriptMessageID);
 
                 newTranscriptMessage.unstable = text;
                 dispatch(updateTranscriptMessage(transcriptMessageID,
