+++ react/features/transcription/reducer.js
@@ -5,14 +5,12 @@ import {
     REMOVE_TRANSCRIPT_MESSAGE,
     UPDATE_TRANSCRIPT_MESSAGE
 } from './actionTypes';
-import { getUpdatedTranscriptionParagraphs } from './functions';
 
 /**
  * Default State for 'features/transcription' feature
  */
 const defaultState = {
-    transcriptMessages: new Map(),
-    transcriptionSubtitles: []
+    transcriptMessages: new Map()
 };
 
 /**
@@ -46,7 +44,7 @@ ReducerRegistry.register('features/transcription', (
  */
 function _addTranscriptMessage(state,
         { transcriptMessageID, participantName }) {
-    const newTranscriptMessages = state.transcriptMessages;
+    const newTranscriptMessages = new Map(state.transcriptMessages);
 
     // Adds a new key,value pair to the Map once a new message arrives.
     newTranscriptMessages.set(transcriptMessageID, { participantName });
@@ -67,16 +65,14 @@ function _addTranscriptMessage(state,
  * reduction of the specified action.
  */
 function _removeTranscriptMessage(state, { transcriptMessageID }) {
-    const newTranscriptMessages = state.transcriptMessages;
+    const newTranscriptMessages = new Map(state.transcriptMessages);
 
     // Deletes the key from Map once a final message arrives.
     newTranscriptMessages.delete(transcriptMessageID);
 
     return {
         ...state,
-        transcriptMessages: newTranscriptMessages,
-        transcriptionSubtitles:
-            getUpdatedTranscriptionParagraphs(newTranscriptMessages)
+        transcriptMessages: newTranscriptMessages
     };
 }
 
@@ -91,15 +87,13 @@ function _removeTranscriptMessage(state, { transcriptMessageID }) {
  */
 function _updateTranscriptMessage(state,
         { transcriptMessageID, newTranscriptMessage }) {
-    const newTranscriptMessages = state.transcriptMessages;
+    const newTranscriptMessages = new Map(state.transcriptMessages);
 
     // Updates the new message for the given key in the Map.
     newTranscriptMessages.set(transcriptMessageID, newTranscriptMessage);
 
     return {
         ...state,
-        transcriptMessages: newTranscriptMessages,
-        transcriptionSubtitles:
-            getUpdatedTranscriptionParagraphs(newTranscriptMessages)
+        transcriptMessages: newTranscriptMessages
     };
 }
